data-platform-agent/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
langchain>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
requests
python-dotenv
